refactor(web): rename retention-policies flag helper off lib/feature-flags
main's sonner-refactor merge also brought in @lorekit/feature-flags
(OpenFeature) with its own packages/web/src/lib/feature-flags/
directory. Renaming this branch's unrelated single-purpose
NEXT_PUBLIC_LOREKIT_RETENTION_POLICIES_ENABLED helper to
lib/retention-policies-flag.ts avoids sitting next to that directory
under a near-identical name. No behavior change.
